backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.27
pydantic[email]>=2.6
email-validator>=2.1
python-docx>=1.1.0
